Roll bindgen to 0.53 and clang-sys to 0.29
This prevents a problem picking up LLVM_CONFIG_PATH on MacOS
with LLVM13. Unclear if this also affects other build configs.

+++ c2rust-ast-exporter/src/lib.rs
@@ -115,7 +115,7 @@ unsafe fn marshal_result(result: *const ExportResult) -> HashMap<String, Vec<u8>
         // Convert CBOR bytes
         let csize = *res.sizes.offset(i);
         let cbytes = *res.bytes.offset(i);
-        let bytes = slice::from_raw_parts(cbytes, csize);
+        let bytes = slice::from_raw_parts(cbytes, csize as usize);
         let mut v = Vec::new();
         v.extend_from_slice(bytes);
